general clean up + remove  SGP4 related functions and redundant class attributes
