+++ src/data/prediction_csv/sim_settings.json
@@ -6,8 +6,8 @@
   "satellite_failure" : 0,
   "espace": false,
   "sim_start_date":"2019-01-01",
-  "sim_end_date":"2023-01-01",
-  "output_frequency": 30,
+  "sim_end_date":"2019-01-07",
+  "output_frequency": 1,
   "sim_object_type": "all",
   "sim_propagation_type": "sgp4",
   "sim_object_catalogue" : "all",
