+++ src/data/results/SATCAT_after_prop_original_aleph_1.csv
@@ -1,60 +1,2 @@
 N,i,h,_soname,_owner,_application,_mass,_maneuverable,_propulsion,_length,_area,total_mass,total_cost
-6372,87.9,1200,Oneweb-2,OneWeb,Comms,200.0,y,electric,1.52,0.7904,1274400,3440880000.0
-4408,70,570,Starlink-1,Starlink,Comms,260.0,y,electric,1.2,0.6911999999999999,1146080,3094416000.0
-30000,53,350,Starlink-2,Starlink,Comms,260.0,y,electric,2.4,1.3823999999999999,7800000,21060000000.0
-2500,33,535,Starlink-2A,Starlink,Comms,260.0,y,electric,1.2,0.6911999999999999,650000,1755000000.0
-2200,53,525,Starlink-2B,Starlink,Comms,260.0,y,electric,1.2,0.6911999999999999,572000,1544400000.0
-2500,43,530,Starlink-2C,Starlink,Comms,260.0,y,electric,1.2,0.6911999999999999,650000,1755000000.0
-192,97,500,ADA_Space,SSOCubesat,EO,8.0,n,None,0.6,0.6,1536,4147200.0
-784,33,590,Amazon_Kuiper-1,Amazon,Comms,325.0,y,electric,1.2,0.6911999999999999,254800,687960000.0
-1296,42,610,Amazon_Kuiper-2,Amazon,Comms,325.0,y,electric,1.2,0.6911999999999999,421200,1137240000.0
-1156,51.9,630,Amazon_Kuiper-3,Amazon,Comms,325.0,y,electric,1.2,0.6911999999999999,375700,1014390000.0
-18,63.4,30000,Apogee_Networks,Other,Comms,20.0,n,None,0.6,0.6,360,972000.0
-243,53.2,510,AST_&_Science,Other,Comms,1500.0,n,electric,0.8,0.6400000000000001,364500,984150000.0
-2240,97,380,AstraSpace-3-SSO,Astraspace,Internet,50.0,y,electric,0.6,0.6,112000,302400000.0
-40,0,700,AstraSpace-EQ,Astraspace,Internet,50.0,y,electric,0.6,0.6,2000,5400000.0
-4896,30,390,AstraSpace-Low-I,Astraspace,Internet,50.0,y,electric,0.6,0.6,244800,660960000.0
-4148,55,400,AstraSpace-Med-I,Astraspace,Internet,50.0,y,electric,0.6,0.6,207400,559980000.0
-1792,55,700,AstraSpace-MED-I,Astraspace,Internet,50.0,y,electric,0.6,0.6,89600,241920000.0
-504,98,690,AstraSpace-SSO,Astraspace,Internet,50.0,y,electric,0.6,0.6,25200,68040000.0
-198,30,1530,Astrome,Other,Comms,10.0,n,None,0.1,0.1,1980,5346000.0
-12,97.7,620,Aurora_Insight,SSOCubesat,Comms,20.0,y,electric,0.6,0.6,240,648000.0
-30,54,400,Avant_Space,Other,Advertising,50.0,y,electric,1.6,1.6,1500,4050000.0
-50,97.7,600,Axelspace,SSOCubesat,EO,100.0,y,gas,0.1,0.1,5000,13500000.0
-14,53,450,Blacksky,Other,EO,55.0,y,electric,3.2,3.2,770,2079000.0
-600,82.9,670,BoeingLEO-2,Boeing,Comms,200.0,y,y,1.52,0.7904,120000,324000000.0
-1400,54.9,680,BoeingLEO-3,Boeing,Comms,250.0,y,electric,1.2,0.6911999999999999,350000,945000000.0
-1564,37.9,690,BoeingLEO-4,Boeing,Comms,250.0,y,electric,1.2,0.6911999999999999,391000,1055700000.0
-840,37.2,1040,BoeingLEO-5,Boeing,Comms,200.0,y,y,1.52,0.7904,168000,453600000.0
-980,48.8,1070,BoeingLEO-6,Boeing,Comms,200.0,y,y,1.52,0.7904,196000,529200000.0
-15,87.9,10000,BoeingMEO,Boeing,Comms,1500.0,y,gas,1.0,0.28,22500,60750000.0
-100,98,500,Canon,SSOCubesat,EO,35.5,y,electric,0.67,0.1943,3500,9585000.0
-14,45.1,556,Capella_Space_Corp-2,Other,EO,40.0,y,electric,0.8,0.020000000000000004,560,1512000.0
-15,97.5,525,Capella_Space_Corp-SSO,Other,EO,100.0,y,electric,0.8,0.020000000000000004,1500,4050000.0
-20,98,520,Carbon_Mapper,SSOCubesat,EO,30.0,y,None,0.3,0.3,600,1620000.0
-50,90,550,Care_Weather,SSOCubesat,Weather,1.0,y,None,0.1,0.1,50,135000.0
-300,90,535,Chang_Guang,SSOCubesat,EO,150.0,y,gas,0.1,0.005000000000000001,45000,121500000.0
-50,86.4,550,Climavision_,SSOCubesat,EO,40.0,y,electric,0.8,0.2,2000,5400000.0
-30,97,500,ConstellR,Other,EO,150.0,y,electric,2.4,2.4,4500,12150000.0
-60,45,550,Curvalux-1,Cubesat,Comms,150.0,y,gas,1.2,0.6911999999999999,9000,24300000.0
-180,87.9,550,Curvalux-2,Cubesat,Comms,150.0,y,gas,1.2,0.6911999999999999,27000,72900000.0
-10,90,550,Earth_Daily_Analytics,SSOCubesat,EO,20.0,y,None,0.2,0.2,200,540000.0
-15,98,505,Earth-I,SSOCubesat,EO,5.0,y,None,0.1,0.1,75,202500.0
-10,65,35786,EchoStar,Other,Comms,200.0,y,gas,3.2,3.2,2000,5400000.0
-7,98,500,EOS_Agrisat,SSOCubesat,EO for agriculture,178.0,y,electric,1.0,0.5,1246,3364200.0
-35,30,600,Eutelsat,Other,IOT,30.0,y,None,0.6,0.6,1050,2835000.0
-68,6,650,ExactEarth_(operating),Other,EO,6.0,y,None,0.02,0.0004,408,1101600.0
-15,97.5,510,Fleet-Space,SSOCubesat,Comms,50.0,y,electric,1.2,1.2,750,2025000.0
-2000,50,600,Guowang-2,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,500000,1350000000.0
-3600,55,508,Guowang-3,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,900000,2430000000.0
-1728,30,1145,Guowang-4,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,432000,1166400000.0
-1728,40,1145,Guowang-5,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,432000,1166400000.0
-1728,50,1145,Guowang-6,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,432000,1166400000.0
-1728,60,1145,Guowang-7,Guowang,Comms,250.0,y,gas,1.2,0.6911999999999999,432000,1166400000.0
-21,97.7,580,Hawkeye__360,Cubesat,Comms,45.0,y,electric,0.5,0.25,945,2551500.0
-1440,55,1150,Hughes_Network_Systems,Other,Comms,250.0,y,electric,1.2,0.6911999999999999,360000,972000000.0
-791,87.9,8000,Mangata_Networks,Other,Comms,200.0,y,y,1.52,0.7904,158200,427140000.0
-1024,60.7,460,SatRevolution,Other,EO,50.0,y,None,0.6,0.6,51200,138240000.0
-120,98.6,800,Theia,SSOCubesat,EO,4.0,No,None,0.3,0.3,480,1296000.0
-5,97.4,495,Zhuhai,SSOCubesat,EO,90.0,?,?,0.1,0.1,450,1215000.0
-5,97.4,510,Zhuhai,SSOCubesat,EO,90.0,?,?,0.1,0.1,450,1215000.0
+192,97.0,500,ADA_Space,SSOCubesat,EO,8.0,n,None,0.6,0.6,1536,4147200.0
