+++ src/main.py
@@ -4,12 +4,13 @@ import os
 import json
 import pickle
 from tqdm import tqdm
-
-# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))) 
+from tqdm.contrib.concurrent import process_map 
+from multiprocessing import Pool, cpu_count
+import cProfile
 
 from utils.SpaceCatalogue import SpaceCatalogue
 from utils.LaunchModel import Prediction2SpaceObjects
-from utils.coords import utc_to_jd
+from utils.Conversions import utc_to_jd
 
 def get_path(*args):
     return os.path.join(os.getcwd(), *args)
@@ -64,7 +65,7 @@ def run_sim(settings):
             decayed_before_start += 1
     print("# sats decayed before sim start date: ", decayed_before_start)
 
-    for satellite in tqdm(SATCAT.Catalogue):
+    for satellite in tqdm(SATCAT.Catalogue): #tqdm is a progress bar
         satellite.prop_catobject(jd_start = jd_start, jd_stop = jd_stop, step_size = 10, propagator="RK45") # propagate each satellite using sgp4, and output a state vector every timestep
     # Export
     print("Exporting results...")
@@ -74,6 +75,68 @@ def run_sim(settings):
     print(f"Number of Satellites: {len(SATCAT.Catalogue)}")
     print("Simulation Complete")
 
+def propagate_satellite(args):
+    satellite, jd_start, jd_stop = args
+    satellite.prop_catobject(jd_start=jd_start, jd_stop=jd_stop, step_size=20, propagator="RK45") # setting step sizde to 20 for now to try and speed up
+    return satellite
+
+def run_parallel_sim(settings):
+    SATCAT = SpaceCatalogue(settings["sim_object_type"], settings["sim_object_catalogue"], settings["repull_catalogues"])
+    jd_start = float(utc_to_jd(settings["sim_start_date"])[0])
+    jd_stop = float(utc_to_jd(settings["sim_end_date"])[0])
+    scenario_name = settings["scenario_name"]
+
+    if settings["repull_catalogues"] and os.path.exists(get_path('src/data/catalogue/All_catalogue_latest.csv')):
+        if settings["sim_object_type"] == "all":
+            SATCAT.CreateCatalogueAll()
+        elif settings["sim_object_type"] != "debris":  
+            SATCAT.CreateCatalogueActive()
+
+        SATCAT.Catalogue2SpaceObjects()
+        
+        dump_pickle('src/data/catalogue/SATCAT_before_prop.pickle', SATCAT)
+
+    else:
+        SATCAT.Catalogue = load_pickle('src/data/catalogue/SATCAT_before_prop.pickle')
+
+    if settings["environment"] == "development" and os.path.exists(get_path(f'src/data/catalogue/SATCAT_before_prop_{scenario_name}.pickle')):
+        SATCAT = load_pickle(f'src/data/catalogue/SATCAT_before_prop_{scenario_name}.pickle')
+    else:
+        if settings["scenario_name"] != "baseline":
+            print("Creating Launch Model...")
+            launch_file_object = Prediction2SpaceObjects('src/data/prediction_csv/FSP_Predictions.csv', 'src/data/prediction_csv/sim_settings.json')
+            SATCAT.Catalogue.extend(launch_file_object)
+
+    print("Number of Satellites: ", len(SATCAT.Catalogue))
+    print(f"Propagating Satellites and saving state every {settings['output_frequency']} days...")
+
+    decayed_before_start = 0
+    for satellite in SATCAT.Catalogue:
+        if satellite.decay_date < datetime.datetime.strptime(settings["sim_start_date"], '%Y-%m-%d'):
+            SATCAT.Catalogue.remove(satellite)
+            decayed_before_start += 1
+    print("# sats decayed before sim start date: ", decayed_before_start)
+
+    # Propagate satellites in parallel
+    print("Propagating satellites in parallel...")
+    SATCAT.Catalogue = process_map(propagate_satellite, [(satellite, jd_start, jd_stop) for satellite in SATCAT.Catalogue], max_workers=cpu_count())
+
+    print("Exporting results...")
+    dump_pickle(f'src/data/catalogue/prop_{scenario_name}.pickle', SATCAT)
+
+    print(f"Output: {get_path(f'src/data/catalogue/{scenario_name}.pickle')}")
+    print(f"Number of Satellites: {len(SATCAT.Catalogue)}")
+    print("Simulation Complete")
+
 if __name__ == '__main__':
+    ##### Profiling #####
     policy = json.load(open(get_path('src/data/prediction_csv/sim_settings.json'), 'r'))
-    run_sim(policy)
+    profiler = cProfile.Profile()
+    profiler.enable()
+    run_parallel_sim(policy)
+    profiler.disable()
+    profiler.print_stats(sort='time')
+    
+    ##### Run sim #####
+    # policy = json.load(open(get_path('src/data/prediction_csv/sim_settings.json'), 'r'))
+    # run_parallel_sim(policy)
