+++ src/utils/SpaceCatalogue.py
@@ -7,7 +7,7 @@ import pandas as pd
 from dotenv import load_dotenv
 import datetime
 from utils.SpaceObject import SpaceObject
-from utils.coords import tle_parse
+from utils.Conversions import tle_parse
 
 class SpaceCatalogue:
     def __init__(self, sim_object_type, sim_object_catalogue, repull_catalogues):
